Move controller logic into a service
