+++ api/v1/services/index.js
@@ -1,5 +1,6 @@
 module.exports = {
 	AuthService: require('./AuthService'),
+	EcosystemService: require('./EcosystemService'),
 	MailService: require('./MailService'),
 	PermissionService: require('./PermissionService'),
 	ProjectService: require('./ProjectService'),
